WIP: start splitting out web repl
