+++ Dockerfile
@@ -9,10 +9,13 @@ RUN unzip ../$COSMO bin/ape.elf bin/assimilate bin/bash bin/python bin/zip
 RUN sh bin/zip -A --delete bin/python "Lib/site-packages/*"
 RUN mkdir Lib
 COPY scrapscript.py Lib
-COPY .args .
+COPY webrepl.py Lib
+RUN echo "-m\nscrapscript\n..." > .args
 RUN bin/ape.elf bin/python -m compileall Lib
 RUN mv Lib/__pycache__/scrapscript*.pyc Lib/scrapscript.pyc
+RUN mv Lib/__pycache__/webrepl*.pyc Lib/webrepl.pyc
 RUN rm Lib/scrapscript.py
+RUN rm Lib/webrepl.py
 RUN cp bin/python bin/scrapscript.com
 COPY style.css Lib
 COPY repl.html Lib
@@ -20,7 +23,13 @@ RUN sh bin/zip -A -r bin/scrapscript.com Lib .args
 RUN bin/ape.elf bin/assimilate bin/scrapscript.com
 
 # Set up the container
-FROM scratch
+FROM scratch as scrapscript
+COPY --from=build /cosmo/bin/scrapscript.com .
+EXPOSE 8000
+ENTRYPOINT ["./scrapscript.com"]
+CMD ["repl"]
+
+FROM scratch as webrepl
 COPY --from=build /cosmo/bin/scrapscript.com .
 EXPOSE 8000
 ENTRYPOINT ["./scrapscript.com"]
